Fix all compile warnings for all targets (#713)
### What problem does this PR solve?

Fix all compile warnings for all targets

### Type of change

- [x] Refactoring

---------

+++ src/unit_test/storage/wal/wal_replay.cpp
@@ -414,15 +414,15 @@ TEST_F(WalReplayTest, WalReplayAppend) {
         //            UniquePtr<MetaTableState> read_table_meta = MakeUnique<MetaTableState>();
         //
         //            txn->GetMetaTableState(read_table_meta.get(), "default", "tbl4", column_ids);
-        //            EXPECT_EQ(read_table_meta->segment_map_.size(), 1);
+        //            EXPECT_EQ(read_table_meta->segment_map_.size(), 1u);
         //
         //            EXPECT_EQ(read_table_meta->local_blocks_.size(), 0);
-        //            EXPECT_EQ(read_table_meta->segment_map_.size(), 1);
+        //            EXPECT_EQ(read_table_meta->segment_map_.size(), 1u);
         //            for (const auto &segment_pair : read_table_meta->segment_map_) {
         //                EXPECT_EQ(segment_pair.first, 0);
         //                EXPECT_NE(segment_pair.second.segment_entry_, nullptr);
-        //                EXPECT_EQ(segment_pair.second.segment_entry_->block_entries_.size(), 1);
-        //                EXPECT_EQ(segment_pair.second.block_map_.size(), 1);
+        //                EXPECT_EQ(segment_pair.second.segment_entry_->block_entries_.size(), 1u);
+        //                EXPECT_EQ(segment_pair.second.block_map_.size(), 1u);
         //                for (const auto &block_pair : segment_pair.second.block_map_) {
         //                    //                    EXPECT_EQ(block_pair.first, 0);
         //                    EXPECT_NE(block_pair.second.block_entry_, nullptr);
@@ -577,7 +577,7 @@ TEST_F(WalReplayTest, WalReplayImport) {
                 auto column_type1 = block_column_entry0->column_type().get();
                 EXPECT_EQ(column_type1->type(), LogicalType::kTinyInt);
                 SizeT data_type_size = columns_vector[0]->data_type_size_;
-                EXPECT_EQ(data_type_size, 1);
+                EXPECT_EQ(data_type_size, 1u);
                 block_column_entry0->Append(columns_vector[0].get(), 0, 1, buffer_manager);
             }
             {
@@ -585,7 +585,7 @@ TEST_F(WalReplayTest, WalReplayImport) {
                 auto column_type2 = block_column_entry1->column_type().get();
                 EXPECT_EQ(column_type2->type(), LogicalType::kBigInt);
                 SizeT data_type_size = columns_vector[1]->data_type_size_;
-                EXPECT_EQ(data_type_size, 8);
+                EXPECT_EQ(data_type_size, 8u);
                 block_column_entry1->Append(columns_vector[1].get(), 0, 1, buffer_manager);
             }
             {
@@ -593,7 +593,7 @@ TEST_F(WalReplayTest, WalReplayImport) {
                 auto column_type3 = block_column_entry2->column_type().get();
                 EXPECT_EQ(column_type3->type(), LogicalType::kDouble);
                 SizeT data_type_size = columns_vector[2]->data_type_size_;
-                EXPECT_EQ(data_type_size, 8);
+                EXPECT_EQ(data_type_size, 8u);
                 block_column_entry2->Append(columns_vector[2].get(), 0, 1, buffer_manager);
             }
 
@@ -631,10 +631,10 @@ TEST_F(WalReplayTest, WalReplayImport) {
             EXPECT_NE(table_entry, nullptr);
             auto *segment_entry = table_entry->GetSegmentByID(0, begin_ts);
             EXPECT_NE(segment_entry, nullptr);
-            EXPECT_EQ(segment_entry->segment_id(), 0);
+            EXPECT_EQ(segment_entry->segment_id(), 0u);
             auto *block_entry = segment_entry->GetBlockEntryByID(0);
-            EXPECT_EQ(block_entry->block_id(), 0);
-            EXPECT_EQ(block_entry->row_count(), 1);
+            EXPECT_EQ(block_entry->block_id(), 0u);
+            EXPECT_EQ(block_entry->row_count(), 1u);
 
             BlockColumnEntry *column0 = block_entry->GetColumnBlockEntry(0);
             BlockColumnEntry *column1 = block_entry->GetColumnBlockEntry(1);
@@ -665,7 +665,7 @@ TEST_F(WalReplayTest, WalReplayImport) {
 }
 
 TEST_F(WalReplayTest, WalReplayCompact) {
-    int test_segment_n = 2;
+    u64 test_segment_n = 2;
     {
         infinity::GlobalResourceUsage::Init();
         std::shared_ptr<std::string> config_path = nullptr;
@@ -697,7 +697,7 @@ TEST_F(WalReplayTest, WalReplayCompact) {
             txn_mgr->CommitTxn(txn);
         }
 
-        for (int i = 0; i < test_segment_n; ++i) { // add 2 segments
+        for (u64 i = 0; i < test_segment_n; ++i) { // add 2 segments
             auto txn2 = txn_mgr->CreateTxn();
             txn2->Begin();
 
@@ -724,7 +724,7 @@ TEST_F(WalReplayTest, WalReplayCompact) {
                 auto column_type0 = block_column_entry0->column_type().get();
                 EXPECT_EQ(column_type0->type(), LogicalType::kTinyInt);
                 SizeT data_type_size = column_vectors[0]->data_type_size_;
-                EXPECT_EQ(data_type_size, 1);
+                EXPECT_EQ(data_type_size, 1u);
                 block_column_entry0->Append(column_vectors[0].get(), 0, 1, buffer_manager);
                 block_entry->IncreaseRowCount(1);
             }
@@ -769,7 +769,7 @@ TEST_F(WalReplayTest, WalReplayCompact) {
             auto [table_entry, status] = txn->GetTableEntry("default", "tbl1");
             EXPECT_NE(table_entry, nullptr);
 
-            for (int i = 0; i < test_segment_n; ++i) {
+            for (u64 i = 0; i < test_segment_n; ++i) {
                 auto *segment = table_entry->GetSegmentByID(i, begin_ts);
                 EXPECT_NE(segment, nullptr);
                 EXPECT_EQ(segment->status(), SegmentStatus::kDeprecated);
@@ -871,7 +871,7 @@ TEST_F(WalReplayTest, WalReplayCreateIndexIvfFlat) {
             auto table_index_meta = table_entry->index_meta_map()["idx1"].get();
             EXPECT_NE(table_index_meta, nullptr);
             EXPECT_EQ(*table_index_meta->index_name(), "idx1");
-            EXPECT_EQ(table_index_meta->index_entry_list().size(), 1);
+            EXPECT_EQ(table_index_meta->index_entry_list().size(), 1u);
             auto table_index_entry_front = static_cast<TableIndexEntry *>(table_index_meta->index_entry_list().front().get());
             EXPECT_EQ(*table_index_entry_front->index_base()->index_name_, "idx1");
             txn_mgr->CommitTxn(txn);
@@ -975,7 +975,7 @@ TEST_F(WalReplayTest, WalReplayCreateIndexHnsw) {
             auto table_index_meta = table_entry->index_meta_map()["hnsw_index"].get();
             EXPECT_NE(table_index_meta, nullptr);
             EXPECT_EQ(*table_index_meta->index_name(), "hnsw_index");
-            EXPECT_EQ(table_index_meta->index_entry_list().size(), 1);
+            EXPECT_EQ(table_index_meta->index_entry_list().size(), 1u);
             auto table_index_entry_front = static_cast<TableIndexEntry *>(table_index_meta->index_entry_list().front().get());
             EXPECT_EQ(*table_index_entry_front->index_base()->index_name_, "hnsw_index");
             txn_mgr->CommitTxn(txn);
